% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_norm.R
\name{mix_norm}
\alias{mix_norm}
\title{Mixture Normal Distribution}
\usage{
mix_norm(p, mu1, mu2, sigma1, sigma2)
}
\arguments{
\item{p}{numeric, probability that Y is drawn from a normal distribution N(mu1,sigma1)}

\item{mu1}{numeric, mean of first normal distribution}

\item{mu2}{numeric, mean of second normal distribution}

\item{sigma1}{numeric, standard deviation of first normal distribution}

\item{sigma2}{numeric, standard deviation of second normal distribution}
}
\value{

}
\description{
Generate n observations from a mixture of two normal distributions: generate U from a Bernoulli(p) distribution
 and if U=1, then draw Y from a normal distribution with mean mu1 and standard deviation sigma1, or if U=0, draw
 Y from the other normal distribution with mean mu2 and standard deviation sigma2
}
\examples{
mix_norm(0.5,-2,3,2,1.5)
}
