% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_simulator.R
\name{weather_simulator}
\alias{weather_simulator}
\title{Weather Simulation}
\usage{
weather_simulator(w)
}
\arguments{
\item{w}{character; "sunny" or "rainy"; the weather condition today.}
}
\value{

}
\description{
Weather is either sunny or rainy in Richmond. If a day is sunny, P(sunny tomorrow)=0.85.
 If a day is rainy, P(rainy tomorrow)=0.35. If a day is rainy, the amount of rainfall accumulation
 in the city is governed by an Exponential(L=2) distribution, where the value from that distribution
 is the rainfall in inches. If a day is sunny, there can be no rain.
 Specifically, given an initial day's weather conditions, simulate the following 10 days of weather and calculate the projected rainfall acc
}
\details{
Given an initial day's weather conditions, simulate the following
10 days of weather and calculate the projected rainfall accumulation in inches.
}
\examples{
weather_simulator("sunny")
}
